# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Personal Finance & Accounting App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary (balances, income, expenses)
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
        - name: currency
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/chart:
    get:
      operationId: getDashboardChart
      tags: [dashboard]
      summary: Get income/expense chart data
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
        - name: currency
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartData"
  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get recent transactions for dashboard
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: object
                required: [transactions]
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
  /dashboard/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [dashboard]
      summary: Get spending breakdown by category
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
            default: expense
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryBreakdown"

  # ─── ACCOUNTS ───────────────────────────────────────────────
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, all]
            default: all
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: object
                required: [categories]
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create custom category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete custom category
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── CURRENCIES ──────────────────────────────────────────────
...
  /currencies/exchange-rates:
    get:
      operationId: getExchangeRates
      tags: [currencies]
      summary: Get exchange rates for a base currency
      security:
        - bearerAuth: []
      parameters:
        - name: base
          in: query
          required: true
          schema:
            type: string
            example: USD
      responses:
        "200":
          description: Exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRates"
  /currencies/convert:
    get:
      operationId: convertCurrency
      tags: [currencies]
      summary: Convert amount between currencies
      security:
        - bearerAuth: []
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionResult"

  # ─── BUDGETS ─────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNotifications"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/read-all:
    put:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/unread-count:
    get:
      operationId: getUnreadNotificationsCount
      tags: [notifications]
      summary: Get unread notifications count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: number
  /notifications/fcm-token:
    put:
      operationId: updateFcmToken
      tags: [notifications]
      summary: Update Firebase Cloud Messaging token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFcmTokenInput"
      responses:
        "200":
          description: Token updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── BACKUPS ─────────────────────────────────────────────────
...
  /admin/currencies:
    get:
      operationId: getAdminCurrencies
      tags: [admin]
      summary: Get all currencies (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Currencies
          content:
            application/json:
              schema:
                type: object
                required: [currencies]
                properties:
                  currencies:
                    type: array
                    items:
                      $ref: "#/components/schemas/Currency"
    post:
      operationId: addCurrency
      tags: [admin]
      summary: Add a currency
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCurrencyRequest"
      responses:
        "201":
          description: Currency added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Currency"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    UpdateFcmTokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
        platform:
          type: string
          enum: [android, ios, web]
...
    DashboardSummary:
      type: object
      required: [totalBalance, totalIncome, totalExpense, netBalance, currency, accountCount, transactionCount]
      properties:
        totalBalance:
          type: number
        totalIncome:
          type: number
        totalExpense:
          type: number
        netBalance:
          type: number
        currency:
          type: string
        accountCount:
          type: number
        transactionCount:
          type: number
        incomeChange:
          type: number
          description: Percentage change from previous period
        expenseChange:
          type: number
          description: Percentage change from previous period
    ChartData:
      type: object
      required: [labels, income, expenses]
      properties:
        labels:
          type: array
          items:
            type: string
        income:
          type: array
          items:
            type: number
        expenses:
          type: array
          items:
            type: number
        currency:
          type: string
    CategoryBreakdown:
      type: object
      required: [items, total, currency]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdownItem"
        total:
          type: number
        currency:
          type: string
    CategoryBreakdownItem:
      type: object
      required: [categoryId, categoryName, categoryIcon, categoryColor, amount, percentage]
      properties:
        categoryId:
          type: string
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        amount:
          type: number
        percentage:
          type: number

    # ─── ACCOUNTS ───────────────────────────────────────────────
    Account:
      type: object
      required: [id, name, type, currencyCode, balance, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, card, savings, investment, other]
        currencyCode:
          type: string
        balance:
          type: number
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        accountNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Transaction:
      type: object
      required: [id, type, amount, currencyCode, accountId, categoryId, date, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        amount:
          type: number
        currencyCode:
          type: string
        exchangeRate:
          type: number
          nullable: true
        accountId:
          type: string
        account:
          $ref: "#/components/schemas/Account"
          nullable: true
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        description:
          type: string
          nullable: true
        receiptUrl:
          type: string
          nullable: true
        transferToAccountId:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Category:
      type: object
      required: [id, name, type, icon, color, isDefault]
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        type:
          type: string
          enum: [income, expense, both]
        icon:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        parentId:
          type: string
          nullable: true
    CreateCategoryRequest:
      type: object
      required: [name, type, icon, color]
      properties:
        name:
          type: string
        nameAr:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        icon:
          type: string
        color:
          type: string
        parentId:
          type: string
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        color:
          type: string

    # ─── CURRENCIES ───────────────────────────────────────────────
    Currency:
      type: object
      required: [id, code, name, symbol, isActive]
      properties:
        id:
          type: string
        code:
          type: string
          example: USD
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        symbol:
          type: string
        flag:
          type: string
          nullable: true
        rateToUsd:
          type: number
          nullable: true
        isActive:
          type: boolean
        updatedAt:
          type: string
          format: date-time
          nullable: true
    ExchangeRates:
      type: object
      required: [base, rates, updatedAt]
      properties:
        base:
          type: string
        rates:
          type: object
          additionalProperties:
            type: number
        updatedAt:
          type: string
          format: date-time
    ConversionResult:
      type: object
      required: [from, to, amount, convertedAmount, rate]
      properties:
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        convertedAmount:
          type: number
        rate:
          type: number
    CreateCurrencyRequest:
      type: object
      required: [code, name, symbol]
      properties:
        code:
          type: string
        name:
          type: string
        nameAr:
          type: string
        symbol:
          type: string
        flag:
          type: string
        rateToUsd:
          type: number

    # ─── BUDGETS ─────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, title, body, type, isRead, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [budget_alert, subscription, system, transaction, backup]
        isRead:
          type: boolean
        data:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaginatedNotifications:
      type: object
      required: [notifications, total, page, limit, totalPages, unreadCount]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
        unreadCount:
          type: number
...